### 2025-05-09 #705 @mpadge
@rmgpanw You've used the pre-submission template here, instead of the one for full submissions. Can you please:

1. click "New Issue" -> "Submit software for review" and copy the templated text somewhere
2. Fill fields by copying across what you've got in current pre-submission template
3. Edit your initial comment by replacing current version will equivalent full-submission version.

The bot checks may not be automatically triggered, so you might also need to manually call `@ropensci-review-bot check package` (as a separate issue comment). Thanks!
