### 2025-05-09 969f5d1
fix test mocking so tests pass when TRUD API key is unavailable
