### 2025-05-14 #705 @mpadge
Thanks for submitting @rmgpanw, and great to see that you've got the tests working. I note only one aspect of the {pkgcheck} results here:

> This package features some noteworthy statistical properties which may need to be clarified by a handling editor prior to progressing.

And clicking on the _statistical properties_ section shows that is just that you have no documentation lines at all for non-exported functions. Documenting all functions both exported and internal greatly helps others to understand, and thereby contribute, to your package. You just need to add a `# @noRd` tag to the end of internal function docs to prevent {roxygen2} from outputting documentation for those. See [this section of our _Dev Guide_](https://devguide.ropensci.org/pkg_building.html#roxygen-2-use) for more details. We'll proceed with review regardless, but do note that documentation of internal functions is also very helpful for reviewers, so please keep both the yet-to-be-assigned editor, and subsequent reviewers informed of your intentions and status in that regard.

Note also that the _statistical properties_ provide immediate context for your statement that,
> `trud` is a relatively small package

In Lines-of-Code, the standard and most direct measure of package size, it is in the 30th percentile, which is below average but definitely not too small, so rest assured, size is not an issue here. But thanks for pointing that out to us!

Other than that, everything has been previously checked and approved in the pre-submission issue #689, so we're good to proceed here. We'll try to find a handling editor asap. You'll see here when that's been arranged.
