### 2025-05-14 f48ca4d
add roxygen documentation for non-exported functions
