### 2025-05-14 #705 @rmgpanw
Thank you for the helpful feedback, as always @mpadge. I have now added some documentation for the non-exported functions in commit [f48ca4d](https://github.com/rmgpanw/trud/commit/f48ca4d16141064a68bfdf54f9b86f3ce0a6690e). Looking forward to making any further improvements!
