+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.2
 pkgdown_sha: ~
 articles:
   trud: trud.html
-last_built: 2025-05-09T08:26Z
+last_built: 2025-05-14T14:01Z
 urls:
   reference: https://rmgpanw.github.io/trud/reference
   article: https://rmgpanw.github.io/trud/articles
