### 2025-05-26 4104d0c
add rOpenSci status badge to README
