+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 3.1.11
-pkgdown: 2.1.2
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   trud: trud.html
-last_built: 2025-05-14T14:01Z
+last_built: 2025-05-26T20:42Z
 urls:
   reference: https://rmgpanw.github.io/trud/reference
   article: https://rmgpanw.github.io/trud/articles
