### 2025-06-17 #705 @jonclayden
---

Thank you for the opportunity to review this package, and for providing an interface to an NHS resource that I was not previously aware of! The basic functionality all seems to work as intended. The API itself is very limited, so I recognise that imposes significant constraints on what you can do, but I hope the following comments and suggestions are helpful.

1. The concept of subscriptions to items is unlikely to be intuitive to a new user. I think it should be mentioned, in the vignette, in the README and in `?trud_items`, that users must subscribe to an item through the NHS TRUD website before they can access it; otherwise they will bump into errors and not know why, or how to resolve them. Perhaps `get_subscribed_metadata()` is a better gateway (at least in some cases) than `trud_items()`, since it only shows subscribed items?
2. I think a little more help is needed in walking the user through the workflow. The vignette shows all the functions in use but provides very little in the way of commentary; the README has a bit more context but doesn't take you all the way through. I would suggest elaborating a little on both so that the process is clearer, and also adding hints to the function documentation, e.g. by saying in `?trud_items` that the `item_number`s in the result can be used as arguments to `get_item_metadata()` and `download_item()`. I'd also suggest using the "see also" Rd field.
3. The package name `trud` should be consistently surrounded by backticks in the README. The font switch is a useful cue that you're referring to the package, not the service it talks to.
4. There's a typo ("publickKey") in the documentation for `download_item()`.
5. Non-obvious processing of the output of `get_item_metadata()`, viz. `purrr::map_at("releases", \(release) purrr::map(release, names))`, is used repeatedly in the examples and vignette, presumably to make the output more readable. I think it should probably be clarified what this is for and, if the API reponse format is reasonably stable, the structure of the return value should be documented in more detail. As in point 2 above, the relevant label to pass as the `release` argument to `download_item()` isn't clear.
6. Since the release names/IDs are a key piece of information, perhaps a variant form (`get_item_releases()`?) could download the metadata and then extract just these IDs?
7. The `TRUD_API_KEY` argument to most functions is counterintuitive: I would expect it to set the key for that call, but in fact it names the environment variable containing it. This strikes me as catering to quite a niche use-case, and since the vignette's instructions say `TRUD_API_KEY` (specifically) should be set in the environment I'm not sure I see the need for it.
8. With no API key set the tests pass with six skips, but with a valid API key set I see two failures (full output below). I think these are expected—one because I'm not subscribed to hardcoded item 394, and one because an expired key is deliberately set—but they're not ideal and potentially misleading. Perhaps the skip checks could be made more robust to avoid them.

library(trud)
testthat::test_local()
## ✔ |          9 | utils
##
## Duration: 5.9 s
##
##  17.             └─rlang::abort(...)
##
##  16.           └─rlang::abort(...)
##
## [ FAIL 2 | WARN 0 | SKIP 0 | PASS 28 ]
##
## Error: Test failures
```
