### 2025-06-18 #705 @AleKoure
- [x] **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports`, and `Maintainer` (which may be autogenerated via `Authors@R`).

---

Great job! 🚀

The `trud` package is a useful tool for accessing NHS TRUD resources, filling a real need for programmatic access. Most functions work as expected, and the API integration is solid. Noteworthy is the inclusion of a scraper to list all available items, a pragmatic and well-executed solution in the absence of a public API endpoint.

Here are a few suggestions to help refine the package:

1. Consider following a consistent style guide (e.g., tidyverse style guide) and adding a lint check in continuous integration. {`lintr`} would be a great fit and help keep the codebase clean and consistent.

2. I suggest using more descriptive text in test descriptions to improve readability and maintainability. For example, in this [test](https://github.com/rmgpanw/trud/blob/4104d0c9fe0a5cfdaa537f8d5c5a873d9c1b327c/tests/testthat/test-download_item.R#L47), the label "download_item() works" could be replaced with something more specific, such as "download_item() downloads a file and throws a warning if the file already exists".

3. `get_trud_api_key("my_key")` throws an error, blocking users from adding the API key directly in the function parameters. Unit test this function and ensure that it behaves as expected when an API key is provided directly as a @param.

4. Regarding this @param [here](https://github.com/rmgpanw/trud/blob/4104d0c9fe0a5cfdaa537f8d5c5a873d9c1b327c/R/get_item_metadata.R#L26):
    - I would suggest using lower case for the parameter name, as it is more common in R packages.
    - Consider instead of using `NULL` as default value, to use `TRUD_API_KEY = Sys.getenv("TRUD_API_KEY")`, as it might be more informative.

5. I suggest using enums (character arguments with `rlang::arg_match`) instead of booleans where possible. For example [here](https://github.com/rmgpanw/trud/blob/4104d0c9fe0a5cfdaa537f8d5c5a873d9c1b327c/R/get_item_metadata.R#L27), replacing `latest_only = FALSE` with `release_scope = c("latest", "all")`. This has the benefit of making function calls more expressive and readable (`release_scope = "all"` is clearer than `latest_only = FALSE`), avoids boolean ambiguity, and is more easily extensible if additional options are introduced in the future. This also simplifies validation, as `rlang::arg_match()` will throw an error if an invalid value is provided.

6. In the description of `get_item_metadata()`, consider noting that users must subscribe to an item before its metadata can be retrieved. This isn’t immediately obvious and could help avoid confusion. A brief mention in the vignette would also be helpful.

7. Similar to point 5 above, this @param [here](https://github.com/rmgpanw/trud/blob/4104d0c9fe0a5cfdaa537f8d5c5a873d9c1b327c/R/download_item.R#L43) can take all available values `file_type = c("archive", "checksum", "signature", "publicKey")` and then use   `file_type <- rlang::arg_match(file_type)` to validate the input and set first element as the default. That way, options are more explicit and validation is simplified.

8. The response object (`resp`) returned [here](https://github.com/rmgpanw/trud/blob/4104d0c9fe0a5cfdaa537f8d5c5a873d9c1b327c/R/download_item.R#L126) is currently unused. Consider extracting and returning the file path from the response `body` itself, as this is more authoritative than constructing the path from the input parameters.

    ```r
    resp <- request_download_item(url, file_path) |>
        purrr::pluck("body") |>
        unclass()

    invisible(resp)
    ```
    Fix any tests that might be affected by this change.

9. In the `download_item()` function, download is skipped if the file already exists. This is a good practice, but consider adding a @param allowing overwriting existing files if needed. Also, [here](https://github.com/rmgpanw/trud/blob/4104d0c9fe0a5cfdaa537f8d5c5a873d9c1b327c/R/download_item.R#L106) consider stating explicitly in the warning that the file already exists and is not being overwritten, to avoid confusion.

    ```r
    c(
        "!" = "Skipping download:",
        "!" = "File {.code {file_name}} already exists in directory {.code {directory}}",
        "i" = "Returning file path {.path {file_path}}"
    )
    ```

10. Consider adding an option or environment variable to allow users to override the default user agent header, and also consider using this mechanism in the CI if applicable. [here](https://github.com/rmgpanw/trud/commit/7c90fdd63abb02ec829fdb78d475c384bea7c513) is a simplistic example of how to do this.

11. Consider adding a request handler for errors, rate limits, and retry logic. This will improve the robustness of the package when dealing with API errors or rate limits.

    example where we set 5 transactions per second and retry up to 3 times on 429 or 503 errors:

    ```
    handle_trud_request <- function(
        req,Add commentMore actions
        max_tries = 3,
        retry_status = c(429, 503),
        capacity = 5,
        fill_time_s = 1
    ) {
        req |>
            httr2::req_error(body = trud_error_message) |>
            httr2::req_retry(
                max_tries = max_tries,
                is_transient = function(resp) {
                    httr2::resp_status(resp) %in% retry_status
                }
            ) |>
            httr2::req_throttle(capacity = capacity, fill_time_s = fill_time_s)
    }
    ```

    Then use this in your API requests constructors, e.g., in `request_download_item()` and `request_item_metadata`

    You can find some information about NFS API standards here: https://digital.nhs.uk/developer/guides-and-documentation/reference-guide

12. This returned value [here](https://github.com/rmgpanw/trud/blob/4104d0c9fe0a5cfdaa537f8d5c5a873d9c1b327c/R/download_item.R#L111) maybe is better to be `invisible()` for consistency.

13. When running tests repeatedly in a local environment, some tests return warnings due to leftover files in `tempdir()`.

This happens because `tempdir()` is shared across test runs and not automatically cleaned between them. To avoid cluttering the file system and prevent such warnings, it's good practice to use `withr::local_tempdir()` instead. This creates an isolated temporary directory for each test and ensures it is cleaned up automatically after execution.

    see example [here](https://github.com/rmgpanw/trud/commit/ff570e0d2843e50f0cba49a81d25f2fbd8d1fc6c)

14. As `trud_items()` is a scraper, add a validation test to ensure that the scraper is still working as expected. This will help catch any changes in the TRUD website structure that might break the scraper in the future.

    ```r
    if (nrow(df) == 0) warning("No TRUD items found — page structure may have changed.")
    ```

15. Consider adding a note in the README for developers to subscribe to [this TRUD category](https://isd.digital.nhs.uk/trud/users/authenticated/filters/0/categories/42) to ensure tests run successfully.

I hope a few of these points help polish things up! 😄
