### 2025-07-24 1a07950
implement error handling and retry logic for TRUD API requests; update tests for expired API key handling
