+++ R/utils.R
@@ -199,3 +199,67 @@ req_user_agent_trud <- function(req) {
   req |>
     httr2::req_user_agent("trud (http://github.com/rmgpanw/trud)")
 }
+
+#' Configure error handling, retry logic, and rate limiting for TRUD API requests
+#'
+#' This function adds request handling to httr2 request objects for
+#' interaction with the NHS TRUD API. It configures custom error messages,
+#' automatic retry for transient failures, and rate limiting to respect API limits.
+#' 
+#' **Retryable errors (will retry automatically):**
+#' - `429` (TOO_MANY_REQUESTS) - Rate limiting (essential for NHS APIs)
+#' - `500` (INTERNAL_SERVER_ERROR) - Server errors that might resolve on retry
+#' - `502` (BAD_GATEWAY) - Gateway/proxy issues that are often temporary  
+#' - `503` (SERVICE_UNAVAILABLE) - Server maintenance/overload
+#' - `504` (GATEWAY_TIMEOUT) - Timeout issues that may succeed on retry
+#'
+#' **Non-retryable errors (will fail immediately):**
+#' - 4xx client errors (400, 401, 403, 404, 409) - Problems with the request
+#' - `501` (NOT_IMPLEMENTED) - The endpoint doesn't support the method
+#'
+#' For further information about NHS API standards, see the 
+#' [NHS reference guide for API standards](https://digital.nhs.uk/developer/guides-and-documentation/reference-guide#api-status).
+#'
+#' @param req An `httr2_request` object created by [httr2::request()].
+#' @param max_tries Integer. Maximum number of retry attempts (default: 3).
+#' @param retry_status An integer vector. HTTP status codes that should trigger
+#'   automatic retry (default: `c(429, 500, 502, 503, 504)`).
+#' @param capacity Integer. Maximum number of requests per `fill_time_s` 
+#'   period for rate limiting (default: 2).
+#' @param fill_time_s Integer Time period in seconds for the rate limiting 
+#'   bucket to refill (default: 2).
+#'
+#' @returns An `httr2_request` object with error handling, retry logic, and 
+#'   rate limiting configured.
+#' @noRd 
+#' @examples
+#' \dontrun{
+#' # Configure a TRUD API request with default retry and rate limiting
+#' req <- httr2::request("https://isd.digital.nhs.uk/trud/api/v1/keys/...")
+#' req <- handle_trud_request(req)
+#' 
+#' # Customize retry behavior
+#' req <- handle_trud_request(
+#'   req, 
+#'   max_tries = 5, 
+#'   capacity = 1, 
+#'   fill_time_s = 3
+#' )
+#' }
+handle_trud_request <- function(
+    req,
+    max_tries = 3,
+    retry_status = c(429, 500, 502, 503, 504),
+    capacity = 2,
+    fill_time_s = 2
+) {
+  req |>
+    httr2::req_error(body = trud_error_message) |>
+    httr2::req_retry(
+      max_tries = max_tries,
+      is_transient = function(resp) {
+        httr2::resp_status(resp) %in% retry_status
+      }
+    ) |>
+    httr2::req_throttle(capacity = capacity, fill_time_s = fill_time_s)
+}
