### 2025-07-24 15b5891
update `download_item()` and `get_item_metadata()` tests requiring a valid API key and subscription to TRUD item 394 - skip these tests if a valid API key is set but without a subscription to item 394
