### 2025-07-24 42c9d5e
update test messages to clarify subscription requirements for TRUD item 394; include subscription link in skip messages
