### 2025-07-24 fe552dd
reformat code using Air and add GitHub Actions workflows for Air (see https://posit-dev.github.io/air/integration-github-actions.html)
