+++ tests/testthat/test-trud_items.R
@@ -1,5 +1,4 @@
 test_that("`trud_items()` returns a tibble with expected properties", {
-
   with_mocked_bindings(
     get_trud_items_html = function() rvest::read_html(snapshot_trud_items_html),
     code = {
@@ -7,9 +6,7 @@ test_that("`trud_items()` returns a tibble with expected properties", {
     }
   )
 
-  expect_true(inherits(trud_items_df,
-    what = "tbl_df"
-  ))
+  expect_true(inherits(trud_items_df, what = "tbl_df"))
 
   expect_true(identical(names(trud_items_df), c("item_number", "item_name")))
 
@@ -26,8 +23,10 @@ test_that("`trud_items()` returns a tibble with expected properties", {
   expect_identical(
     trud_items_df,
     trud_items_df |>
-      dplyr::filter(!is.na(item_number) &
-        !is.na(item_name))
+      dplyr::filter(
+        !is.na(item_number) &
+          !is.na(item_name)
+      )
   )
 })
 
