+++ R/trud_items.R
@@ -37,8 +37,13 @@ trud_items <- function() {
 
   # Filter for items
   df |>
-    dplyr::filter(stringr::str_detect(.data[["item_link"]], "trud/users/guest/filters/0/categories/1/items")) |>
-    dplyr::filter(!.data[["item_name"]] %in% c("Releases", "Licences", "Future releases")) |>
+    dplyr::filter(stringr::str_detect(
+      .data[["item_link"]],
+      "trud/users/guest/filters/0/categories/1/items"
+    )) |>
+    dplyr::filter(
+      !.data[["item_name"]] %in% c("Releases", "Licences", "Future releases")
+    ) |>
     dplyr::select(dplyr::all_of(c("item_number", "item_name")))
 }
 
@@ -47,5 +52,7 @@ trud_items <- function() {
 #' @return Parsed html.
 #' @noRd
 get_trud_items_html <- function() {
-  rvest::read_html("https://isd.digital.nhs.uk/trud/users/guest/filters/0/categories/1")
+  rvest::read_html(
+    "https://isd.digital.nhs.uk/trud/users/guest/filters/0/categories/1"
+  )
 }
