### 2025-07-24 98e937f
Remove TRUD_API_KEY argument from functions and update documentation and tests accordingly
