+++ tests/testthat/test-get_item_metadata.R
@@ -56,22 +56,22 @@ test_that("get_item_metadata() constructs correct URLs whether retrieving for al
   skip_if(condition = identical(Sys.getenv("TRUD_API_KEY"), ""))
   skip_if(condition = identical(Sys.getenv("PKG_CHECK"), "true")) # see pkgcheck.yaml
   # latest_only = TRUE
-  withr::with_envvar(
-    new = c("EXPIRED_API_KEY" = "e963cc518cc41500e1a8940a93ffc3c0915e2983"),
+  tryCatch(
     {
-      tryCatch(
+      withr::with_envvar(
+        c("TRUD_API_KEY" = "e963cc518cc41500e1a8940a93ffc3c0915e2983"),
         {
           metadata_394_lastet_only <- get_item_metadata(
             394,
-            TRUD_API_KEY = "EXPIRED_API_KEY",
             latest_only = TRUE
           )
-        },
-        error = \(cnd) invisible()
+        }
       )
-      req <- httr2::last_request()
-    }
+    },
+    error = \(cnd) invisible()
   )
+
+  req <- httr2::last_request()
 
   expect_equal(
     req$url,
@@ -84,10 +84,14 @@ test_that("get_item_metadata() constructs correct URLs whether retrieving for al
     {
       tryCatch(
         {
-          metadata_394_lastet_only <- get_item_metadata(
-            394,
-            TRUD_API_KEY = "EXPIRED_API_KEY",
-            latest_only = FALSE
+          withr::with_envvar(
+            c("TRUD_API_KEY" = "e963cc518cc41500e1a8940a93ffc3c0915e2983"),
+            {
+              metadata_394_lastet_only <- get_item_metadata(
+                394,
+                latest_only = FALSE
+              )
+            }
           )
         },
         error = \(cnd) invisible()
@@ -108,9 +112,9 @@ test_that("get_item_metadata() throws unauthorized error when API key is expired
   skip_if(condition = identical(Sys.getenv("PKG_CHECK"), "true")) # see pkgcheck.yaml
   expect_error(
     withr::with_envvar(
-      new = c("EXPIRED_API_KEY" = "e963cc518cc41500e1a8940a93ffc3c0915e2983"),
+      new = c("TRUD_API_KEY" = "e963cc518cc41500e1a8940a93ffc3c0915e2983"),
       {
-        get_item_metadata(1799, TRUD_API_KEY = "EXPIRED_API_KEY")
+        get_item_metadata(1799)
       }
     ),
     "UNAUTHORIZED: API access is disabled for the requesting account."
