+++ R/get_item_metadata.R
@@ -21,10 +21,10 @@
 #'
 #' @examples
 #' # An informative error is raised if your API key is invalid or missing
-#' try(download_item(394, TRUD_API_KEY = "INVALID_API_KEY"))
-get_item_metadata <- function(item, TRUD_API_KEY = NULL, latest_only = FALSE) {
+#' try(withr::with_envvar(c("TRUD_API_KEY" = ""), get_item_metadata(394)))
+get_item_metadata <- function(item, latest_only = FALSE) {
   # validate args
-  TRUD_API_KEY <- get_trud_api_key(TRUD_API_KEY)
+  TRUD_API_KEY <- get_trud_api_key()
 
   validate_arg_item(item = item)
 
