### 2025-07-26 0c1c13a
update documentation - clarify need for users to subscribe to TRUD items before they can access them
