+++ R/get_subscribed_metadata.R
@@ -1,16 +1,18 @@
 #' Get metadata for subscribed NHS TRUD items
 #'
 #' A convenience wrapper around [trud_items()] and [get_item_metadata()],
-#' retrieving metadata for only items that the user is subscribed to.
+#' retrieving metadata for only items that the user is subscribed to. If you
+#' need access to additional items, browse available options with
+#' [trud_items()], then subscribe through the NHS TRUD website.
 #'
 #' @inheritParams get_item_metadata
 #'
 #' @returns A tibble, with item metadata stored in the list column `metadata`.
 #' @export
-#' @seealso  [trud_items()], [get_item_metadata()]
+#' @seealso [trud_items()] to browse all available items, [get_item_metadata()]
 #'
 #' @examplesIf identical(Sys.getenv("IN_PKGDOWN"), "true")
-#' get_subscribed_metadata()
+#'   get_subscribed_metadata()
 get_subscribed_metadata <- function(latest_only = FALSE) {
   # validate args
   get_trud_api_key()
