+++ man/trud_items.Rd
@@ -12,9 +12,20 @@ A tibble, with columns \code{item_number} and \code{item_name}.
 \description{
 Scrapes \href{https://isd.digital.nhs.uk/trud/users/guest/filters/0/categories/1}{this page}
 from the NHS TRUD website for all available items.
+
+\emph{\strong{Subscription Required}}
+
+You must subscribe to TRUD items individually through the NHS TRUD website
+before you can access them using \code{\link[=get_item_metadata]{get_item_metadata()}} or \code{\link[=download_item]{download_item()}}.
+Simply having an API key is not sufficient. To see items you're already
+subscribed to, use \code{\link[=get_subscribed_metadata]{get_subscribed_metadata()}}. To browse all available
+items, use \code{\link[=trud_items]{trud_items()}}.
 }
 \examples{
 \dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 trud_items()
 \dontshow{\}) # examplesIf}
+}
+\seealso{
+\code{\link[=get_subscribed_metadata]{get_subscribed_metadata()}} to see only your subscribed items.
 }
