### 2025-07-27 0681c62
Enhance documentation for NHS TRUD functions to clarify item retrieval and subscription requirements
