### 2025-07-27 6cd9021
explain use of purrr::map_at() in examples for get_item_metadata() to avoid showing sensitive API keys in URLs
