### 2025-07-27 62f1db9
clarify how to obtain release IDs from `get_item_metadata()` to download specific releases with `download_item()`
