### 2025-07-27 6bc131f
Refactor `get_item_metadata()` and `download_item()` to use `release_scope` parameter instead of `latest_only`
