### 2025-07-27 50944af
Update `download_item()` to replace `download_file` parameter with `file_type`, and use rlang::arg_match for validation.
