### 2025-07-27 8ddee29
update download_item() to return file path from the response body itself, rather than constructing the path from the input parameters.
