### 2025-07-27 6d795ef
Allow users to override the default user agent header by setting the `TRUD_USER_AGENT` environment variable.
