### 2025-07-27 cb46399
Consistently return file path invisibly from `download_item()` function
