### 2025-07-27 df5ef55
Refactor tests to use `withr::local_tempdir()` instead of `tempdir()`.
