### 2025-07-27 3fd3efa
Update NEWS.md
