### 2025-07-27 042abc3
remove non-ASCII characters; update WORDLIST; reformat code using Air
