+++ R/trud_items.R
@@ -4,16 +4,16 @@
 #' Scrapes [this
 #' page](https://isd.digital.nhs.uk/trud/users/guest/filters/0/categories/1)
 #' from the NHS TRUD website for all available items. The `item_number` column
-#' in the result contains the identifiers you need for [get_item_metadata()] 
+#' in the result contains the identifiers you need for [get_item_metadata()]
 #' and [download_item()].
 #'
 #' ```{r child = "man/rmd/subscription-required.Rmd"}
 #' ```
 #'
-#' @returns A tibble, with columns `item_number` and `item_name`. Use the 
+#' @returns A tibble, with columns `item_number` and `item_name`. Use the
 #'   `item_number` values as arguments to [get_item_metadata()] and [download_item()].
 #' @export
-#' @seealso 
+#' @seealso
 #' * [get_subscribed_metadata()] to see only items you're subscribed to
 #' * [get_item_metadata()] to get detailed information about a specific item
 #' * [download_item()] to download files for a specific item
@@ -56,23 +56,23 @@ trud_items <- function() {
       !.data[["item_name"]] %in% c("Releases", "Licences", "Future releases")
     ) |>
     dplyr::select(dplyr::all_of(c("item_number", "item_name")))
-  
+
   # Validate scraper is working
   if (nrow(result) == 0) {
     cli::cli_warn(c(
-      "!" = "No TRUD items found — page structure may have changed.",
+      "!" = "No TRUD items found - page structure may have changed.",
       "i" = "This may indicate changes to the NHS TRUD website that require package updates.",
       "i" = "Please report this issue at: {.url https://github.com/rmgpanw/trud/issues}"
     ))
   } else if (nrow(result) < 10) {
     cli::cli_warn(c(
-      "!" = "Unusually few TRUD items found ({nrow(result)} items) — page structure may have partially changed.",
+      "!" = "Unusually few TRUD items found ({nrow(result)} items) - page structure may have partially changed.",
       "i" = "Expected to find many more items on the NHS TRUD website.",
       "i" = "This may indicate parsing issues that require package updates.",
       "i" = "Please report this issue at: {.url https://github.com/rmgpanw/trud/issues}"
     ))
   }
-  
+
   result
 }
 
