### 2025-07-31 #705 @AleKoure
## Reviewer Response

Great work! 🚀
On my side, everything looks solid functionally and structurally.

Two optional notes for future updates:

**3 & 4.** Relying only on env vars for API keys may be limiting. Passing them as args can help in testing or multi-key use.

**5.** `rlang::arg_match()` can validate choices, no need for indirection (see also [here](https://design.tidyverse.org/boolean-strategies.html)):

```r
release_scope <- rlang::arg_match(release_scope)
```

**+** For API testing, consider looking into **{webfakes}** and **{httrtest2}**.

#### Final approval (post-review)

Estimated hours spent reviewing: 2
