### 2025-07-31 #705 @rmgpanw
Great news and super helpful feedback as always @AleKoure, I am very grateful :) I had not come across https://design.tidyverse.org/ or the 2 packages you mentioned for API testing, so am looking forward to learning more. Thank you!
