### 2025-08-06 #705 @jonclayden
## Reviewer Response

Well done on the updates, @rmgpanw – I think they really help the user walk through usage of the package and get the most from it. I no longer see test failures, and the skip message is very clear.

My only very minor thought would be to add a step 0 (or new step 1) to the very helpful "recommended workflow" section in the README: namely, to ensure the API key is set. I know it's mentioned just above that section, but I think these things are worth saying repeatedly, especially since nothing very useful can be done without that step, and it makes the workflow complete. Your call, though.

#### Final approval (post-review)

Estimated hours spent reviewing: 1
