### 2025-08-06 38265f2
Add 'step 0' (setting TRUD_API_KEY in .Renviron) to recommended workflow in README
