+++ README.Rmd
@@ -58,6 +58,14 @@ You will also need to [sign up for a free account](https://isd.digital.nhs.uk/tr
 
 ### Recommended Workflow
 
+**Step 0:**: Set up your TRUD API key as an environmental variable named `TRUD_API_KEY`. For example, create or edit your project `.Renviron` file with `usethis::edit_r_environ()`, then populate as follows (replacing `e963cc518cc41500e1a8940a93ffc3c0915e2983` with your own API key):[^1]
+
+[^1]: You will also need to restart your R session to set any environmental variables that have been newly-added to your project `.Renviron` file.
+
+```
+TRUD_API_KEY=e963cc518cc41500e1a8940a93ffc3c0915e2983
+```
+
 **Step 1**: Check what you're already subscribed to:
 
 ```{r subscribed-items}
