### 2025-08-08 #705 @ropensci-review-bot
Approved! Thanks @rmgpanw for submitting and @AleKoure, @jonclayden for your reviews! :grin:
