### 2025-08-09 #705 @rmgpanw
Brilliant news! Thank you @emilyriederer :)

@jonclayden and @AleKoure I have now added you both as reviewers in the DESCRIPTION file (https://github.com/ropensci/trud/commit/ad9c855b8fd4cf60d4e17a9597c1d6fe256b34cf). Please let me know if you would like me to amend anything, otherwise I will go ahead with submitting the updated package version to CRAN in the next 2-3 days. Thanks again!
